% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appt_last.R
\name{appt_last}
\alias{appt_last}
\title{appt_last}
\usage{
appt_last(c_date = Sys.Date(), d_pills, days = NULL)
}
\arguments{
\item{c_date}{ISO8601}

\item{d_pills}{integer duration of pills dispensed}

\item{days}{integer, number of days of appointment}
}
\value{
long date formate
}
\description{
appt_last
}
\examples{
pill_days <- 7
appt_last(d_pills = pill_days)
}
