% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appt_next.R
\name{appt_next}
\alias{appt_next}
\title{appt_next}
\usage{
appt_next(c_date = Sys.Date(), d_pills, days = NULL)
}
\arguments{
\item{c_date}{ISO8601}

\item{d_pills}{integer duration of pills dispensed}

\item{days}{integer, number of days of appointment}
}
\value{
nicely formatted date
}
\description{
appt_next
}
\examples{
appt_next(d_pills = 14)
}
